requirements.txt
fastmcp
mcp
markdownify
readabilipy
httpx
pydantic
openai
python-docx
PyPDF2
python-dotenv